Edit recent projects style, also move the file
