+++ src/css/styles.css
@@ -123,25 +123,27 @@ main modal. They basically override bootstrap styling for these components
   font-weight: bold;
 }
 
-#table td:nth-child(n+3):nth-child(-n+5) {
+#recentProjects td:nth-child(n+3):nth-child(-n+5) {
   font-style: italic;
+  width: 100px
 }
-#table td:nth-child(2) {
+#recentProjects td:nth-child(2) {
     font-weight: bold;
+    width: 100px
 }
 
-#table>tbody>tr>td {
+#recentProjects>tbody>tr>td {
   padding-top: 0;
   padding-bottom: 0;
 }
-#table td:nth-child(n+1):nth-child(-n+5) {
+#recentProjects td:nth-child(n+1):nth-child(-n+5) {
   padding-top: 15px;
 }
-#table th {
+#recentProjects th {
   font-weight: normal;
 }
 
-#table>tbody>tr {
+#recentProjects>tbody>tr {
     border-bottom: 1px solid var(--border-color);
 }
